refactor: rename app to Omi Offline and fix critical stability issues
App:
- Renamed package to com.omi.offline and updated all platform metadata.
- Fixed Red Screen crashes by adding missing internationalization support to MaterialApp.
- Implemented comprehensive null-safety refactor across DeviceProvider, DeviceSettings, and transport layers.
- Fixed SD card sync logic to ensure stable recording IDs and visibility of unsynced folders.

Firmware:
- Fixed compilation errors related to current_mtu and Kconfig dependencies.
- Fixed critical storage sync bug by correcting GATT attribute indices and implementing MTU-aware packet sizing.
- Cleaned up numerous compiler warnings across battery, storage, and imu modules.
- Incremented firmware version to oo-1.0.4.

+++ omi/firmware/omi/src/lib/core/utils.h
@@ -11,8 +11,8 @@
     }
 
 #define ASSERT_TRUE(result)                                                                                            \
-    if (!result) {                                                                                                     \
-        LOG_ERR("Error at %s:%d:%d", __FILE__, __LINE__, result);                                                      \
+    if (!(result)) {                                                                                                   \
+        LOG_ERR("Error at %s:%d:%d", __FILE__, __LINE__, (int)(result));                                               \
         return -1;                                                                                                     \
     }
